+++ test/device_marketing_names_test.dart
@@ -18,6 +18,18 @@ void main() {
 
   TestWidgetsFlutterBinding.ensureInitialized();
 
+  const _model = "SM-J250F";
+  const _modelFirstResult = "Galaxy Grand Prime Pro";
+  const _modelResult = "Galaxy Grand Prime Pro / Galaxy J2 / Galaxy J2 Pro";
+  const _modelExampleList = "some / device / data";
+  const _modelExampleSome = "some";
+  const _modelExampleDevice = "device";
+  const _modelExampleData = "data";
+  const _modelExampleInvalid = "invalid_data";
+  const _modelExampleEmpty = "";
+  const _modelIos = "iPhone13,4";
+  const _modelIosResult = "iPhone 12 Pro Max";
+
   group('Core functions', () {
     test('Lookup Browser', () async {
       when(platform.isWeb()).thenReturn(true);
@@ -33,11 +45,11 @@ void main() {
       when(platform.isWeb()).thenReturn(false);
       when(platform.isAndroid()).thenReturn(true);
       when(device.getAndroidInfo())
-          .thenAnswer((_) async => getMockAndroidInfo("KB2005"));
+          .thenAnswer((_) async => getMockAndroidInfo(_model));
 
       final result = await lookupDevice(platform, device);
 
-      expect(result, "OnePlus 8T");
+      expect(result, _modelResult);
     });
 
     test('Lookup iOS', () async {
@@ -45,155 +57,190 @@ void main() {
       when(platform.isAndroid()).thenReturn(false);
       when(platform.isIOS()).thenReturn(true);
       when(device.getIosInfo())
-          .thenAnswer((_) async => getMockIosInfo("iPhone13,4"));
+          .thenAnswer((_) async => getMockIosInfo(_modelIos));
 
       final result = await lookupDevice(platform, device);
 
-      expect(result, "iPhone 12 Pro Max");
+      expect(result, _modelIosResult);
     });
 
-    test('Lookup iOS by model', () {
+    test('Lookup names by Android model', () {
       final deviceNames = DeviceMarketingNames();
-      final result =
-          deviceNames.getMarketingNameFromModel(DeviceType.ios, "iPhone13,4");
+      final result = deviceNames.getNamesFromModel(DeviceType.android, _model);
 
-      expect(result, 'iPhone 12 Pro Max');
+      expect(result, _modelResult);
     });
 
-    test('Lookup Android by model', () {
+    test('Lookup names by Android model as list', () {
       final deviceNames = DeviceMarketingNames();
       final result =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "KB2005");
+          deviceNames.getNamesFromModelAsList(DeviceType.android, _model);
 
-      expect(result, 'OnePlus 8T');
+      expect(result.length, 3);
+      expect(result.contains(_modelFirstResult), true);
+    });
+
+    test('Lookup single name by Android model', () {
+      final deviceNames = DeviceMarketingNames();
+      final result =
+          deviceNames.getSingleNameFromModel(DeviceType.android, _model);
+
+      expect(result, _modelFirstResult);
+    });
+
+    test('Lookup single name by Android invalid model', () {
+      final deviceNames = DeviceMarketingNames();
+      final result = deviceNames.getSingleNameFromModel(
+          DeviceType.android, _modelExampleInvalid);
+
+      expect(result, _modelExampleInvalid);
+    });
+
+    test('Lookup names by iOS model', () {
+      final deviceNames = DeviceMarketingNames();
+      final result = deviceNames.getNamesFromModel(DeviceType.ios, _modelIos);
+
+      expect(result, _modelIosResult);
+    });
+
+    test('Lookup names by iOS model as list', () {
+      final deviceNames = DeviceMarketingNames();
+      final result =
+          deviceNames.getNamesFromModelAsList(DeviceType.ios, _modelIos);
+
+      expect(result.length, 1);
+      expect(result.contains(_modelIosResult), true);
+    });
+
+    test('Lookup single name by iOS model', () {
+      final deviceNames = DeviceMarketingNames();
+      final result =
+          deviceNames.getSingleNameFromModel(DeviceType.ios, _modelIos);
+
+      expect(result, _modelIosResult);
+    });
+
+    test('Lookup single name by iOS invalid model', () {
+      final deviceNames = DeviceMarketingNames();
+      final result = deviceNames.getSingleNameFromModel(
+          DeviceType.ios, _modelExampleInvalid);
+
+      expect(result, _modelExampleInvalid);
     });
 
     test('Verify data', () {
       final deviceNames = DeviceMarketingNames();
       final result0 =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "001HT");
+          deviceNames.getNamesFromModel(DeviceType.android, "001HT");
       final result1 =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "10 Plus");
-      final result2 = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "2000_000170_00");
-      final result3 =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "3000");
+          deviceNames.getNamesFromModel(DeviceType.android, "10 Plus");
+      final result2 =
+          deviceNames.getNamesFromModel(DeviceType.android, "2000_000170_00");
+      final result3 = deviceNames.getNamesFromModel(DeviceType.android, "3000");
       final result4 =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "4003J");
+          deviceNames.getNamesFromModel(DeviceType.android, "4003J");
       final result5 =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "5-Inches");
-      final result6 =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "6002");
+          deviceNames.getNamesFromModel(DeviceType.android, "5-Inches");
+      final result6 = deviceNames.getNamesFromModel(DeviceType.android, "6002");
       final result7 =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "701HW");
-      final result8 =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "80-1");
-      final result9 = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "9 inch SDU");
+          deviceNames.getNamesFromModel(DeviceType.android, "701HW");
+      final result8 = deviceNames.getNamesFromModel(DeviceType.android, "80-1");
+      final result9 =
+          deviceNames.getNamesFromModel(DeviceType.android, "9 inch SDU");
       final resultA =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "A-TT00");
-      final resultB = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "B BOT 550");
-      final resultC = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "C Bot Tab 70");
-      final resultD = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "D-805XATV HD");
+          deviceNames.getNamesFromModel(DeviceType.android, "A-TT00");
+      final resultB =
+          deviceNames.getNamesFromModel(DeviceType.android, "B BOT 550");
+      final resultC =
+          deviceNames.getNamesFromModel(DeviceType.android, "C Bot Tab 70");
+      final resultD =
+          deviceNames.getNamesFromModel(DeviceType.android, "D-805XATV HD");
       final resultE =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "E-Tab 4G");
-      final resultF = deviceNames.getMarketingNameFromModel(
+          deviceNames.getNamesFromModel(DeviceType.android, "E-Tab 4G");
+      final resultF = deviceNames.getNamesFromModel(
           DeviceType.android, "F and U Android TV");
-      final resultG = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "G Guard SMART TV");
-      final resultH =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "H-35");
-      final resultI = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "I star two");
-      final resultJ =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "J101");
-      final resultK = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "K-BOOK73G");
+      final resultG =
+          deviceNames.getNamesFromModel(DeviceType.android, "G Guard SMART TV");
+      final resultH = deviceNames.getNamesFromModel(DeviceType.android, "H-35");
+      final resultI =
+          deviceNames.getNamesFromModel(DeviceType.android, "I star two");
+      final resultJ = deviceNames.getNamesFromModel(DeviceType.android, "J101");
+      final resultK =
+          deviceNames.getNamesFromModel(DeviceType.android, "K-BOOK73G");
       final resultL =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "L-01E");
-      final resultM = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "M BOT TAB 100");
+          deviceNames.getNamesFromModel(DeviceType.android, "L-01E");
+      final resultM =
+          deviceNames.getNamesFromModel(DeviceType.android, "M BOT TAB 100");
       final resultN =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "N-01D");
-      final resultO = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "O2 TV Box");
+          deviceNames.getNamesFromModel(DeviceType.android, "N-01D");
+      final resultO =
+          deviceNames.getNamesFromModel(DeviceType.android, "O2 TV Box");
       final resultP =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "P-01K");
-      final resultQ = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "Q Explorist HR");
+          deviceNames.getNamesFromModel(DeviceType.android, "P-01K");
+      final resultQ =
+          deviceNames.getNamesFromModel(DeviceType.android, "Q Explorist HR");
       final resultR =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "R1 HD");
+          deviceNames.getNamesFromModel(DeviceType.android, "R1 HD");
       final resultS =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "S-SAB220");
+          deviceNames.getNamesFromModel(DeviceType.android, "S-SAB220");
       final resultT =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "T-01D");
+          deviceNames.getNamesFromModel(DeviceType.android, "T-01D");
       final resultU =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "U FEEL");
-      final resultV = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "V TAB 10'' LITE");
-      final resultW = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "W-K130-EEA");
+          deviceNames.getNamesFromModel(DeviceType.android, "U FEEL");
+      final resultV =
+          deviceNames.getNamesFromModel(DeviceType.android, "V TAB 10'' LITE");
+      final resultW =
+          deviceNames.getNamesFromModel(DeviceType.android, "W-K130-EEA");
       final resultX =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "X MAX");
-      final resultY =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "Y10");
+          deviceNames.getNamesFromModel(DeviceType.android, "X MAX");
+      final resultY = deviceNames.getNamesFromModel(DeviceType.android, "Y10");
       final resultZ =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "Z-2252");
-      final resulta =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "a4");
+          deviceNames.getNamesFromModel(DeviceType.android, "Z-2252");
+      final resulta = deviceNames.getNamesFromModel(DeviceType.android, "a4");
       final resultb =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "b1-720");
+          deviceNames.getNamesFromModel(DeviceType.android, "b1-720");
       final resultc =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "calgary");
+          deviceNames.getNamesFromModel(DeviceType.android, "calgary");
       final resultd =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "d-01H");
+          deviceNames.getNamesFromModel(DeviceType.android, "d-01H");
       final resulte =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "e-tab 20");
-      final resultf =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "f400");
-      final resultg =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "g06");
-      final resulth =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "h33");
-      final resulti = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "i-STYLE 713");
+          deviceNames.getNamesFromModel(DeviceType.android, "e-tab 20");
+      final resultf = deviceNames.getNamesFromModel(DeviceType.android, "f400");
+      final resultg = deviceNames.getNamesFromModel(DeviceType.android, "g06");
+      final resulth = deviceNames.getNamesFromModel(DeviceType.android, "h33");
+      final resulti =
+          deviceNames.getNamesFromModel(DeviceType.android, "i-STYLE 713");
       final resultj =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "jupiter");
+          deviceNames.getNamesFromModel(DeviceType.android, "jupiter");
       final resultk =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "keeby");
+          deviceNames.getNamesFromModel(DeviceType.android, "keeby");
       final resultl =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "l4300");
+          deviceNames.getNamesFromModel(DeviceType.android, "l4300");
       final resultm =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "mDESK");
-      final resultn = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "nJoy_Arcas_7");
-      final resulto = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "omega_lite_4");
+          deviceNames.getNamesFromModel(DeviceType.android, "mDESK");
+      final resultn =
+          deviceNames.getNamesFromModel(DeviceType.android, "nJoy_Arcas_7");
+      final resulto =
+          deviceNames.getNamesFromModel(DeviceType.android, "omega_lite_4");
       final resultp =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "pearl");
+          deviceNames.getNamesFromModel(DeviceType.android, "pearl");
       final resultq =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "quattro");
-      final resultr = deviceNames.getMarketingNameFromModel(
+          deviceNames.getNamesFromModel(DeviceType.android, "quattro");
+      final resultr = deviceNames.getNamesFromModel(
           DeviceType.android, "realme SA Smart TV");
-      final results =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "s4");
+      final results = deviceNames.getNamesFromModel(DeviceType.android, "s4");
       final resultt =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "tab10");
-      final resultu = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "uie4027lgu");
+          deviceNames.getNamesFromModel(DeviceType.android, "tab10");
+      final resultu =
+          deviceNames.getNamesFromModel(DeviceType.android, "uie4027lgu");
       final resultv =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "v7-EEA");
-      final resultw =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "w6");
-      final resultx = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "xTablet-A680");
-      final resulty = deviceNames.getMarketingNameFromModel(
-          DeviceType.android, "yetruepad21");
-      final resultz =
-          deviceNames.getMarketingNameFromModel(DeviceType.android, "zx70");
+          deviceNames.getNamesFromModel(DeviceType.android, "v7-EEA");
+      final resultw = deviceNames.getNamesFromModel(DeviceType.android, "w6");
+      final resultx =
+          deviceNames.getNamesFromModel(DeviceType.android, "xTablet-A680");
+      final resulty =
+          deviceNames.getNamesFromModel(DeviceType.android, "yetruepad21");
+      final resultz = deviceNames.getNamesFromModel(DeviceType.android, "zx70");
 
       expect(result0, 'Desire HD');
       expect(result1, '10 Plus');
@@ -258,18 +305,6 @@ void main() {
       expect(resulty, 'yetruepad21');
       expect(resultz, 'ZX70');
     });
-
-    test('Fail lookup, no match', () {
-      final result = lookupName(DeviceType.ios, "no_match");
-
-      expect(result, null);
-    });
-
-    test('Fail lookup, null', () {
-      final result = lookupName(DeviceType.ios, null);
-
-      expect(result, null);
-    });
   });
 
   group('Utils', () {
@@ -277,6 +312,47 @@ void main() {
       final result = 'a'.sentenceCase();
 
       expect(result, 'A');
+    });
+
+    test('Get list with multiple devices', () {
+      final result = _modelExampleList.getList();
+
+      expect(result.length, 3);
+      expect(true, result.contains(_modelExampleSome));
+      expect(true, result.contains(_modelExampleDevice));
+      expect(true, result.contains(_modelExampleData));
+    });
+
+    test('Get list with single device', () {
+      final result = _modelExampleDevice.getList();
+
+      expect(result.length, 1);
+      expect(true, result.contains(_modelExampleDevice));
+    });
+
+    test('Get list with no devices', () {
+      final result = _modelExampleEmpty.getList();
+
+      expect(result.length, 1);
+      expect(true, result.contains(_modelExampleEmpty));
+    });
+
+    test('Get first entry with multiple devices', () {
+      final result = _modelExampleList.getFirst();
+
+      expect(result, _modelExampleSome);
+    });
+
+    test('Get first entry with single device', () {
+      final result = _modelExampleDevice.getFirst();
+
+      expect(result, _modelExampleDevice);
+    });
+
+    test('Get first entry with no devices', () {
+      final result = _modelExampleEmpty.getFirst();
+
+      expect(result, _modelExampleEmpty);
     });
   });
 }